# backend/docs/swagger.yaml
definitions:
...
  models.ErrorResponse:
    properties:
      '{route} error':
        example: '{method}: {crud}: error message'
        type: string
    type: object
...
  models.UserWeightCreateRequestModel:
    properties:
      weight:
        maximum: 300
        minimum: 10
        type: number
    required:
    - weight
    type: object
  models.UserWeightResponseModel:
    properties:
      date:
        type: string
      id:
        type: string
      weight:
        type: number
    type: object
info:
  contact: {}
paths:
...
  /user/weight:
    get:
      description: Получает текущий вес пользователя
      produces:
      - application/json
      responses:
        "200":
          description: Актуальный (последний) вес пользователя
          schema:
            $ref: '#/definitions/models.UserWeightResponseModel'
        "400":
          description: Неверный формат ID
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Отсутствует авторизация
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Получение текущего веса пользователя
      tags:
      - User Weight
    patch:
      consumes:
      - application/json
      description: Обновляет текущий вес пользователя
      produces:
      - application/json
      responses:
        "200":
          description: Успешное обновление
          schema:
            type: string
        "400":
          description: Ошибка валидации или неверный формат ID
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Отсутствует авторизация
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Обновление веса пользователя (пока нет)
      tags:
      - User Weight
    post:
      consumes:
      - application/json
      description: Создает новую запись о весе пользователя
      parameters:
      - description: Данные веса для создания
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.UserWeightCreateRequestModel'
      produces:
      - application/json
      responses:
        "200":
          description: Успешное создание
          schema:
            type: string
        "400":
          description: Ошибка валидации или неверный формат ID
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Отсутствует авторизация
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Создание записи о весе пользователя
      tags:
      - User Weight
  /user/weight/{uuid}:
    delete:
      description: Удаляет запись о весе пользователя по ID
      parameters:
      - description: ID записи о весе
        in: path
        name: uuid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Успешное удаление
          schema:
            type: string
        "400":
          description: Неверный формат ID
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Отсутствует авторизация
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Удаление записи о весе пользователя
      tags:
      - User Weight
  /user/weight/history:
    get:
      description: Получает всю историю изменений веса пользователя
      produces:
      - application/json
      responses:
        "200":
          description: История веса пользователя
          schema:
            items:
              items:
                $ref: '#/definitions/models.UserWeightResponseModel'
              type: array
            type: array
        "400":
          description: Неверный формат ID
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Отсутствует авторизация
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Получение истории веса пользователя
      tags:
      - User Weight
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...